backend/requirements.txt
cffi==1.14.6
click==8.0.1
cryptography==3.4.8
Flask==2.0.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
Pillow==8.3.2
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
python-decouple==3.4
six==1.16.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
